colocate util function and rename for better understanding
