+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/main/utils/get_ancestors_index_by_id.ts
@@ -7,7 +7,7 @@
 
 import { ALERT_RULE_TYPE } from '@kbn/rule-data-utils';
 import type { TimelineEventsDetailsItem } from '@kbn/timelines-plugin/common';
-import { getClusterQualifiedIndex } from '../../../shared/utils/get_cluster_qualified_index';
+import { getNonLocalQualifiedIndex } from '../../../shared/utils/non_local_index';
 import {
   EVENT_SOURCE_FIELD_NAME,
   LEGACY_EVENT_SOURCE_FIELD_NAME,
@@ -54,7 +54,7 @@ export const getAncestorsIndexById = (
       ids.forEach((id, i) => {
         const indexName = indices[i];
         if (id && indexName) {
-          acc[id] = getClusterQualifiedIndex(indexName, documentIndex);
+          acc[id] = getNonLocalQualifiedIndex(indexName, documentIndex);
         }
       });
       return acc;
