+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/tools/analyzer/index.tsx
@@ -17,7 +17,7 @@ import { useSelectedPatterns } from '../../../../data_view_manager/hooks/use_sel
 import { useDataView } from '../../../../data_view_manager/hooks/use_data_view';
 import { useTimelineDataFilters } from '../../../../timelines/containers/use_timeline_data_filters';
 import { Resolver } from '../../../../resolver/view';
-import { withDocumentIndex } from '../../../../resolver/utils/with_document_index';
+import { withDocumentIndex } from '../../../shared/utils/non_local_index';
 import { ANALYZER_TITLE } from '../../../shared/constants/flyout_titles';
 
 export const ANALYZER_GRAPH_TEST_ID = `${PREFIX}AnalyzerGraph` as const;
