+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/shared/utils/non_local_index.ts
@@ -5,6 +5,20 @@
  * 2.0.
  */
 
+import { isNonLocalIndexName } from '@kbn/es-query';
+
+/**
+ * Qualifies an index name recorded inside a document (e.g. `kibana.alert.ancestors[].index`) with
+ * the cross-cluster / cross-project alias of the document that contains it.
+ */
+export const getNonLocalQualifiedIndex = (index: string, documentIndex: string): string => {
+  if (!index || isNonLocalIndexName(index) || !isNonLocalIndexName(documentIndex)) {
+    return index;
+  }
+  const separatorIndex = documentIndex.indexOf(':');
+  return `${documentIndex.slice(0, separatorIndex + 1)}${index}`;
+};
+
 /**
  * Prepends a project-qualified document `_index` (`alias:index`) so entity lookup can pick the
  * clicked document when the same `_id` exists in more than one project.
@@ -14,7 +28,7 @@
  * as Analyzer's "this alert from being analyzed" error.
  */
 export const withDocumentIndex = (indices: string[], documentIndex?: string | null): string[] => {
-  if (!documentIndex || !documentIndex.includes(':') || indices.includes(documentIndex)) {
+  if (!documentIndex || !isNonLocalIndexName(documentIndex) || indices.includes(documentIndex)) {
     return indices;
   }
